Remove unnecessary auth_fingerprint post param to FPTI url
