+++ Sources/BraintreeCore/Analytics/BTAnalyticsMetadata.swift
