+++ Sources/BraintreeCore/Analytics/Encodable+Dictionary.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+extension Encodable {
+
+    /// Converts to dictionary `[String: Any]` type.
+    ///
+    /// Used to satisfy limitations of current BTHTTP implementation.
+    func toDictionary(_ encoder: JSONEncoder = JSONEncoder()) throws -> [String: Any] {
+        let data = try encoder.encode(self)
+        let object = try JSONSerialization.jsonObject(with: data)
+        guard let json = object as? [String: Any] else {
+            let context = DecodingError.Context(codingPath: [], debugDescription: "Deserialized object is not a dictionary")
+            throw DecodingError.typeMismatch(type(of: object), context)
+        }
+        return json
+    }
+}
