+++ Sources/BraintreeCore/Analytics/FPTIBatchEvent.swift
